[sl-core] Fix input of slr variables.
* tools/bin/sag: Ensure that the data size is populated in the
last entry.

* lib/src/mtinput.c: Simplify.

+++ slc/lib/src/mtinput.c
@@ -37,7 +37,7 @@ void sys_vars_init(void *rawdata, bool copy)
     if (nitems == 0)
         sz = sizeof(slr_base[0]);
     else
-        sz = (size_t)slr_base[nitems].nitems;
+        sz = slr_base[nitems].nitems;
 
     if (copy)
     {
